move user-progress to component folder
