+++ src/app/modules/item/components/user-progress/user-progress.component.scss
